+++ packages/svelte/src/internal/client/runtime.js
@@ -1,5 +1,5 @@
 import { subscribe_to_store } from '../../store/utils.js';
-import { EMPTY_FUNC } from '../common.js';
+import { EMPTY_FUNC } from '../common/index.js';
 import { unwrap } from './render.js';
 import { map_delete, map_get, map_set } from './operations.js';
 import { is_array } from './utils.js';
