+++ packages/svelte/src/internal/server/index.js
@@ -1,6 +1,6 @@
 import * as $ from '../client/runtime.js';
 import { set_is_ssr } from '../client/runtime.js';
-import { is_promise } from '../common.js';
+import { is_promise } from '../common/index.js';
 import { subscribe_to_store } from '../../store/utils.js';
 
 export * from '../client/validate.js';
